pyproject.toml
[tool.poetry]
name = "blog-rest-api"
version = "0.1.0"
description = ""
authors = ["Sulav Maharjan <sulavmaharjan63@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
djoser = "^2.1.0"
django-debug-toolbar = "^3.4.0"
psycopg2-binary = "^2.9.3"
django-cors-headers = "^3.13.0"
Pillow = "^9.1.1"
python-dotenv = "^0.20.0"
django-taggit = "^3.0.0"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
